# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Mimir API","version":"0.1.0"},"paths":{"/metrics":{"get":{"summary":"Metrics","description":"Endpoint that serves Prometheus metrics.","operationId":"metrics_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/healthz":{"get":{"summary":"Healthz","operationId":"healthz_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthStatus"}}}}}}},"/readyz":{"get":{"summary":"Readyz","operationId":"readyz_readyz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthStatus"}}}}}}},"/api/tasks":{"get":{"tags":["tasks"],"summary":"List Tasks","operationId":"list_tasks_api_tasks_get","parameters":[{"name":"horizon","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/Horizon"},{"type":"null"}],"title":"Horizon"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":3,"title":"Limit"}},{"name":"X-Dev-User","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Dev-User"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskList"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tasks/{task_id}/complete":{"post":{"tags":["tasks"],"summary":"Complete Task","operationId":"complete_task_api_tasks__task_id__complete_post","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}},{"name":"X-Dev-User","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Dev-User"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tasks/{task_id}/undo":{"post":{"tags":["tasks"],"summary":"Undo Task","operationId":"undo_task_api_tasks__task_id__undo_post","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}},{"name":"X-Dev-User","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Dev-User"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tasks/recompute_priorities":{"post":{"tags":["tasks"],"summary":"Recompute Priorities","operationId":"recompute_priorities_api_tasks_recompute_priorities_post","parameters":[{"name":"X-Dev-User","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Dev-User"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tasks/suggest":{"post":{"tags":["tasks"],"summary":"Suggest Tasks","operationId":"suggest_tasks_api_tasks_suggest_post","parameters":[{"name":"X-Dev-User","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Dev-User"}}],"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/graph":{"get":{"tags":["graph"],"summary":"Get Graph","operationId":"get_graph_api_graph_get","parameters":[{"name":"window","in":"query","required":false,"schema":{"type":"string","default":"month","title":"Window"}},{"name":"X-Dev-User","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Dev-User"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GraphResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/graph/filters":{"get":{"tags":["graph"],"summary":"Graph Filters","operationId":"graph_filters_api_graph_filters_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/connectors":{"get":{"tags":["connectors"],"summary":"List Connectors","operationId":"list_connectors_api_connectors_get","parameters":[{"name":"X-Dev-User","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Dev-User"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConnectorList"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/connectors/{kind}/connect":{"post":{"tags":["connectors"],"summary":"Start Connect","operationId":"start_connect_api_connectors__kind__connect_post","parameters":[{"name":"kind","in":"path","required":true,"schema":{"type":"string","title":"Kind"}},{"name":"X-Dev-User","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Dev-User"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Start Connect Api Connectors  Kind  Connect Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/connectors/{kind}/authorize":{"get":{"tags":["connectors"],"summary":"Authorize Connector","operationId":"authorize_connector_api_connectors__kind__authorize_get","parameters":[{"name":"kind","in":"path","required":true,"schema":{"type":"string","title":"Kind"}},{"name":"X-Dev-User","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Dev-User"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/connectors/{kind}/test":{"post":{"tags":["connectors"],"summary":"Test Connector","operationId":"test_connector_api_connectors__kind__test_post","parameters":[{"name":"kind","in":"path","required":true,"schema":{"type":"string","title":"Kind"}},{"name":"X-Dev-User","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Dev-User"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConnectorTestResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/connectors/test_all":{"post":{"tags":["connectors"],"summary":"Test All Connectors","operationId":"test_all_connectors_api_connectors_test_all_post","parameters":[{"name":"X-Dev-User","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Dev-User"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/oauth/start/{kind}":{"get":{"tags":["oauth"],"summary":"Oauth Start","operationId":"oauth_start_api_oauth_start__kind__get","parameters":[{"name":"kind","in":"path","required":true,"schema":{"type":"string","title":"Kind"}},{"name":"X-Dev-User","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Dev-User"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/oauth/callback/{kind}":{"get":{"tags":["oauth"],"summary":"Oauth Callback","operationId":"oauth_callback_api_oauth_callback__kind__get","parameters":[{"name":"kind","in":"path","required":true,"schema":{"type":"string","title":"Kind"}},{"name":"code","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Code"}},{"name":"X-Dev-User","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Dev-User"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/dev/seed":{"post":{"tags":["dev"],"summary":"Seed","operationId":"seed_api_dev_seed_post","parameters":[{"name":"X-Dev-User","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Dev-User"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Connector":{"properties":{"id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Id"},"kind":{"type":"string","title":"Kind"},"status":{"type":"string","title":"Status","default":"disconnected"},"message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Message"}},"type":"object","required":["kind"],"title":"Connector"},"ConnectorList":{"properties":{"connectors":{"items":{"$ref":"#/components/schemas/Connector"},"type":"array","title":"Connectors"}},"type":"object","required":["connectors"],"title":"ConnectorList"},"ConnectorTestResult":{"properties":{"kind":{"type":"string","title":"Kind"},"status":{"type":"string","title":"Status"},"message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Message"}},"type":"object","required":["kind","status"],"title":"ConnectorTestResult"},"GraphResponse":{"properties":{"nodes":{"items":{"$ref":"#/components/schemas/Task"},"type":"array","title":"Nodes"},"edges":{"items":{"items":{"type":"string","format":"uuid"},"type":"array"},"type":"array","title":"Edges"},"meta":{"anyOf":[{},{"type":"null"}],"title":"Meta"}},"type":"object","required":["nodes","edges"],"title":"GraphResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthStatus":{"properties":{"status":{"type":"string","title":"Status","example":"ok"}},"type":"object","required":["status"],"title":"HealthStatus"},"Horizon":{"type":"string","enum":["today","week","month","past7d"],"title":"Horizon"},"Status":{"type":"string","enum":["todo","in_progress","done","scheduled"],"title":"Status"},"Task":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"horizon":{"$ref":"#/components/schemas/Horizon"},"status":{"$ref":"#/components/schemas/Status"},"priority":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Priority"}},"type":"object","required":["id","title","horizon","status"],"title":"Task"},"TaskList":{"properties":{"tasks":{"items":{"$ref":"#/components/schemas/Task"},"type":"array","title":"Tasks"}},"type":"object","required":["tasks"],"title":"TaskList"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}